+++ package.json
@@ -1,27 +1,20 @@
 {
   "name": "matrix.me",
   "version": "0.0.1",
-  "private": true,
   "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear --single"
+    "dev": "svelte-kit dev",
+    "build": "svelte-kit build",
+    "package": "svelte-kit package",
+    "preview": "svelte-kit preview",
+    "prepare": "svelte-kit sync"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-node-resolve": "^11.0.0",
-    "rollup": "^2.3.4",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^7.0.0",
-    "rollup-plugin-terser": "^7.0.0",
+    "@sveltejs/adapter-auto": "next",
+    "@sveltejs/kit": "1.0.0-next.308",
     "svelte": "^3.48.0"
   },
+  "type": "module",
   "dependencies": {
-    "bootstrap": "^5.1.3",
-    "node-sass": "^7.0.1",
-    "page": "^1.11.6",
-    "sirv-cli": "^2.0.0",
     "sveltestrap": "^5.9.0"
   }
 }
