--- public/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset='utf-8'>
-	<!-- <meta name='viewport' content='width=device-width,initial-scale=1'> -->
-
-	<title>Matrix.me</title>
-	<link rel='stylesheet' href='css/global.css'>
-	<link rel='icon' type='image/png' href='/favicon.png'>
-	<link rel='stylesheet' href='/build/bundle.css'>
-	<script defer src='/build/bundle.js'></script>
-</head>
-
-<body>
-</body>
-
-</html>
