+++ src/routes/__layout.svelte
@@ -1,0 +1,9 @@
+<script>
+    import Header from "$lib/Header.svelte";
+import "../styles/global.css"
+</script>
+<header>
+    <Header />
+</header>
+<slot></slot>
+
