+++ src/styles/global.css
@@ -1,0 +1,9 @@
+@import "https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css";
+@import url('https://fonts.googleapis.com/css2?family=Koulen&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Koulen&family=Rajdhani:wght@300;400&display=swap');
+
+body {
+    margin: 0;
+    padding: 0;
+    background-color: #282a36;
+}
