+++ svelte.config.js
@@ -1,0 +1,16 @@
+import adapter from '@sveltejs/adapter-auto';
+
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
+	kit: {
+		adapter: adapter(),
+		vite: {
+			optimizeDeps: {
+			  entries: []
+			}
+		  }
+
+	}
+};
+
+export default config;
